resolved issues and refactored TurnManager
