+++ src/services/turnManager/index.ts
@@ -1,2 +1,4 @@
 export { TurnManager } from './TurnManager';
+export { TurnSystemService } from './TurnSystemService';
+export type { TurnState } from './TurnManager';
